[WIP] Moved FilterValuesBag creating code from the formatter to Input/Query
The formatter had a big design flaw, that it was only capable of processing a QueryFilter string.
Requiring all inputs to return this format, even tho they already had a valid FilterValuesBag object.

An input class is now responsible for knowing what field-names are accepted, and holding the configuration.
The class will throw an exception when an required field is missing, or an none accepted value is encountered. It does not validate the input in depth, the formatter is responsible for this.

Last, the hasGroups() is removed. And getValues() is renamed to getGroups()