[Doctrine] added an conversion strategy pattern for a more flexible conversion
This allows for using an different conversion 'strategy' depending on the current value.

For example: The user want to work with an birthday and actual age.
Depending on the current value, the db-value must be converted to an age
and else be used as it is.

Each conversion strategy will use a different 'slot' in the query
like if the user was using an separate field.