Refactored MessageBag and FilterTypeInterface
An filter-type no longer uses the $message parameter but uses $messageBag (MessageBag instance).
And based on the existence of error messages the validation state is determined.
This change makes it more in-line with the Validator component.

MessageBag API has changed addError() and addInfo() is now always translated.
The $addTranslatorPrefix parameter is removed, and added support for plural messages.

And last, the default domain for addError() is changed to 'validators'.