[Type] removed fall-back support for big numbers
This was originally added for providing big number support, but because of the
complexity of some locals its to difficult actually get this right always.
So instead its better to use a 64bit version of PHP when working with big numbers.

The support is not completely removed. Its just that parsing and formatting of
big (international) numbers beyond the maximum integer bit length will fail.

For the cases of big numbers that are not localized, bcmath will be used when
possible.