[Doctrine] fixed circular reference issue with var_export() and conversions
The Doctrine ORM query cache uses var_export() to make a cache key, and as the WhereBuilder
uses query hints for conversions making it refer to its self. But the WhereBuilder
has a reference to the container and Doctrine this all explodes.

So instead we pass a closure with a reference to the WhereBuilder to the QueryHint.
And that works perfect.