added new Doctrine/Orm component
The QueryWhereBuilderHelper is removed because a WhereBuilder instance
is now linked to the Query and SearchCondition

Registering conversions is now only possible explicitly (no more crazy mapping hacks)

Special conversion strategy 0 is removed as it was to much of an implementation detail
and did not work with DQL, but the actual usage is still possible using a SqlValueConversion.
The same technique is also used for Regex pattern-matching.