mark SearchConditionBuilder:__construct protected
Usage of this method is protected is now considered and access
will be changed to protected in 2.0. The static create() method should
be used instead.

Creating a SearchCondition without a FieldSet is impossible
but the constructor can't enforce this as its also used internally
creating a reference for each group!