Introduce FieldSetConfigurator for keeping config
Instead of Creating a FieldSet directly or storing it in the app-config,
the FieldSet “configuration” is kept in a Configurator (PHP class).

The SearchFactory::createFieldSet() tries to find the FieldSet “configuration” from a registry.
Note that FieldSetRegistryInterface no longer allows to get all FieldSet (configurations) in a single
call, as configurators can also be loaded using there FQCN.

The created FieldSet has the class name of the Configurator for validation reference.

Note: This implementation is inspired on the Rollerworks Datagrid component, but
doesn’t have options for the configurator as this breaks with serialization auto-loading.