Rename some classes to remove interfaces suffix
This commit prepares the code base to allow removing the Interface suffix,
as renaming SearchFactoryInterface to SearchFactory conflicts with the existing
class.

This is a two stage process to not make the diff’s to messy.

—
The (new) style guide of Rollerworks forbids that Interfaces are suffixed
with “Interface” as this breaks ambiguous naming.

Say we had the SearchFactory class but like to introduce a SearchFactory (Interface),
renaming the class is one thing. But now all the typehints and references need to be updated,
which can cause some serieus BC breakage.

Second Interfaces suffixes are a form Hungarian notations, which is considered a bad practice
for a really long time now. Only traits should be suffixed as they can never change there intention,
not can they be type-hinted.