Introduce a new Lexer for QueryStringInput
The original Lexer used the Doctrine Lexer which turned each character
in a token-id. But for the input system that was overly complicated
and made "loos" input impossible.

The new Lexer uses a cursor concept that allows to change the expectations
while parsing. Instead of running trough the data as whole each part is
analyzed and acted on accordingly. As a result the following is now possible:

* All characters (except special syntax characters) can now be used without
  surrounding them with quotes. $12.00 is now possible!

* Line numbers are now properly reported, and the column position is more
  accurate. At best this is reported at the actual character position.

* Incorrectly escaped values will now give a friendly error message:
  * 'A value containing spaces must be surrounded by quotes'
  * 'A value containing special characters must be surrounded by quotes'
  * 'A value containing special characters must be surrounded by quotes'

**Breaking changes:**

* Spaces are more strict now.
* Spaces are no longer allowed between operators, `~ >` this is invalid now.
* A value cannot contain new lines, this was already never supported
  but was never properly reported.
* New lines are only allowed _after_ a value separator or FieldIdentifier,
  between group braces, and before a semi colon `;`.
* The Range value now uses `~` instead `-`.

And as a big bonus the Lexer can be reused for the newly planned SmartQuery
input processor!