Simplify column/value conversion system
Value conversions was introduced when prepared statement parameters were still supported.
This functionality however has been removed as it caused to much issues, but the conversion system itself
was never updated to this new design.

The ValueConversionInterface and ValueConversionInterface are redundant in there purpose, and have been merged
together. A ValueConversion is now expected to return a properly formatted and escaped SQL statement.

* Rename SqlFieldConversionInterface to ColumnConversion
* Rename SqlValueConversion to ValueConversion
* Rename ConversionStrategyInterface to StrategySupportedConversion