(Re) add Doctrine Configuration handling
The Doctrine Configuration handling was moved to it’s own bundles but the maintenance cost
in comparison to the benefits is minimum. So to make installation friendlier and maintenance easier
this will register and configure RollerworksSearch Doctrine DBAL/ORM extension automatically when possible.
Following the same approach as the SearchProcessor.