Redesign the Metadata structure for better DX
Instead of using the FieldSet as configuration reference the metadata now works with contexts.
Context `_any` is used when no context is provided by a custom listener.

A context is resolved in the SearchConditionListener and passed trough the system using
the `_api_search_config` Request attribute. Condition processors no longer require the MetadataFactory.

Secondly a `_default` context is introduced that allows to define defaults for all contexts, the configuration
of `_default` is merged by the DefaultConfigurationMetadataFactory. So runtime overhad is minimum.

Last, the error reporting is greatly improved and the URL generator will now include all previous _route_params
unless they begin with an underscore or param value is an array.

FieldSet building using the Metadata is removed, this would be to limited and was removed from the Core
library for this exact reason. Also it would only require more maintenance and increase complexity.