[Doctrine] Fix Cache key ignoring precondition
The Doctrine DBAL/ORM Condition Caches did not include the pre-condition when computing the cache-key.

The PreCondition is part of the generated Query but the cache key doesn't change for a different PreCondition.
If a non-restricting PreCondition was used before, the second cache hit will use this, ignoring the PreCondition with restrictions!

In practice leading to an information disclosure.